New allocation method for Entity System
Instead of treating each chunk as an entry in a linked list that needs to be traversed every time an entity is allocated or freed, we store a bit array next to the entity system.

The bit array contains one bit for every chunk. It is MUCH master to check if 4 chunks are available using a singe bitmask, rather than having to jump through hoops with memory.

Since we no longer have to traverse the entity system itself, we also don't need to store pointers to the first available entities of each size anymore. This means we don't have to keep those updated, every time we do any single thing.

This SIGNIFICANTLY speeds up allocating and ESPECIALLY freeing entities. In a system with 64 chunks, worst-case allocation takes between 1 and 2 scanlines of CPU time, and worst case freeing takes 68 cycles (+ overhead for call/ret instructions).